backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-cors==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
flask-bcrypt==1.0.1
PyJWT==2.7.0
python-jose==3.3.0
email-validator==2.0.0